requirements.txt
tensorflow-gpu==1.8
fire
tqdm
h5py
keras
numpy
six
mmh3
konlpy
pandas